Renamed ODE_and_signal_functions and moved functions over to analysis.py
