+++ analysis.py
@@ -2,7 +2,29 @@ from parameters import *
 import numpy as np
 from scipy.optimize import fsolve
 from scipy.integrate import odeint
-from ODE_and_Signal_functions import myofib_macro_ODE_reverse
+
+
+#Outputs list of gradients, state encompasses concentrations for mF, M, CSF and PDGF as cells per ml
+def myofib_macro(state, t): # outputs list of gradients
+    
+    mF, M, CSF, PDGF = state
+
+    d_mF_dt = mF * (lambda1 * (PDGF / (k1 + PDGF)) * (1 - mF / K) - mu1)
+    d_M_dt = M * (lambda2 * (CSF / (k2+CSF)) - mu2)
+    d_CSF_dt = beta1 * mF - alpha1 * M * (CSF / (k2 + CSF)) - gamma * CSF
+    d_PDGF_dt = beta2 * M + beta3 * mF - alpha2 * mF * (PDGF / (k1 + PDGF))- gamma * PDGF
+
+    return [d_mF_dt, d_M_dt, d_CSF_dt, d_PDGF_dt]
+
+
+
+#outputs reverse derivative, state encompasses mF, M concentrations 
+def myofib_macro_ODE_reverse(state, t):
+    derivatives = myofib_macro(state, t)
+    
+    #checking if concentrations are in a reasonable range
+    return [-d for d in derivatives] if all(0 <= x <= 10**7 for x in state) else [0, 0]
+
 
 #find steady state for CSF and PDGF given mF and M levels using the fast timescale 
 def CSF_PDGF_steady(x):
@@ -31,6 +53,10 @@ def mF_M_rates(x, t):
     d_mF_dt = mF * (lambda1 * ((PDGF)/(k1+PDGF))*(1-mF/K)-mu1)
     d_M_dt = M*(lambda2*(CSF/(k2 + CSF))- mu2)
     return [d_mF_dt, d_M_dt]
+
+def rev_mf_M_rates(state, t):
+    derivatives = mF_M_rates(state, t)
+    return [-d for d in derivatives]
 
 
 def CSF_PDGF_steady_array(x): # finds steady CSF and PDGF levels for given mF and M levels
